Share parser for parse specs. Speeds up tests enough to be worth the dirtiness I feel inside for doing it.
The main slowness of the test suite was constantly reading the regexes
from disk and parsing them for each new spec.

This makes it so the parser loads the patterns once and parses them
with yaml once for all the #parse specs.

On my machine, the test suite went from like 2 minutes to 7 seconds.
*DRAMATIC HAMSTER*